fix a number of minor issues in the state view:
1. peak and rate computations were using uint32 size which caused
the fields from working for states that transferred more than 4GB
also fix an unnecessary cast to double found while looking for the bug.

2. When creating new cache entries for existing states, start time from
0 rather than the current time to prevent a "spike" in rate and peak.

+++ src/usr.bin/systat/cache.h
@@ -30,7 +30,7 @@ struct sc_ent {
 	double		    peak;
 	double		    rate;
 	time_t		    t;
-	u_int32_t	    bytes;
+	u_int64_t	    bytes;
 };
 
 int cache_init(int);
